+++ petitRADTRANS/__init__.py
@@ -1,9 +1,5 @@
-from __future__ import print_function
-
 import os
 import sys
-
-from petitRADTRANS.config import petitradtrans_config
 
 # Link to the libs folders on Windows
 extra_dll_dir = os.path.join(os.path.dirname(__file__), '.libs')