+++ petitRADTRANS/ccf/ccf_utils.py
@@ -6,11 +6,13 @@ from petitRADTRANS.ccf.etc_cli_module import *
 from petitRADTRANS.ccf.mock_observation import *
 from petitRADTRANS.ccf.model_containers import *
 
+from petitRADTRANS import phoenix
+
 
 def calculate_star_snr(wavelengths, star_effective_temperature, star_radius, star_distance, exposure_time,
                        telescope_mirror_radius, telescope_throughput, instrument_resolving_power,
                        pixel_per_resolution_element=2):
-    stellar_spectral_radiance = nc.get_PHOENIX_spec(star_effective_temperature)
+    stellar_spectral_radiance = phoenix.get_PHOENIX_spec(star_effective_temperature)
     wavelength_stellar = stellar_spectral_radiance[:, 0]  # in cm
 
     wh = np.where(np.logical_and(
@@ -35,7 +37,7 @@ def calculate_star_snr(wavelengths, star_effective_temperature, star_radius, sta
 
 
 def calculate_star_radiosity(wavelength_boundaries, star_effective_temperature, star_radius, star_distance):
-    stellar_spectral_radiance = nc.get_PHOENIX_spec(star_effective_temperature)
+    stellar_spectral_radiance = phoenix.get_PHOENIX_spec(star_effective_temperature)
     wavelength_stellar = stellar_spectral_radiance[:, 0]  # in cm
     stellar_spectral_radiance = SpectralModel.radiosity_erg_hz2radiosity_erg_cm(
         stellar_spectral_radiance[:, 1],