+++ petitRADTRANS/poor_mans_nonequ_chem/poor_mans_nonequ_chem.py
@@ -8,7 +8,7 @@ import h5py
 import os
 from .chem_fortran_util import chem_fortran_util as cfu
 import copy as cp
-from petitRADTRANS import petitradtrans_config
+from petitRADTRANS.config import petitradtrans_config
 
 path = petitradtrans_config['Paths']['pRT_input_data_path']