+++ petitRADTRANS/retrieval/examples/transmission/master_retrieval_model.py
@@ -3,9 +3,9 @@
 ##----------------------------------------
 
 import numpy as np
-import sys
-from petitRADTRANS import Radtrans
+
 from petitRADTRANS import nat_cst as nc
+from petitRADTRANS import physics
 
 
 def calc_MMW(abundances):
@@ -45,7 +45,7 @@ def retrieval_model_plain(rt_object, temperature_parameters, log_g, log_P0, \
     gravity = 1e1 ** log_g
 
     # Create temperature model
-    press, temp = nc.make_press_temp(temperature_parameters)  # pressures from low to high
+    press, temp = physics.make_press_temp(temperature_parameters)  # pressures from low to high
 
     abundances = {}
     metal_sum = 0.