+++ tests/context.py
@@ -7,6 +7,8 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')
 import petitRADTRANS
 import petitRADTRANS.fort_rebin
 import petitRADTRANS.nat_cst
+import petitRADTRANS.phoenix
+import petitRADTRANS.physics
 import petitRADTRANS.poor_mans_nonequ_chem
 import petitRADTRANS.radtrans
 import petitRADTRANS.version